Moves auth routes to seperate file
